+++ plugins/writeopia_persistence_core/src/commonMain/kotlin/io/writeopia/sdk/persistence/core/repository/InMemoryDocumentRepository.kt
@@ -86,7 +86,7 @@ class InMemoryDocumentRepository : DocumentRepository {
         }
     }
 
-    override suspend fun saveStoryStep(storyStep: StoryStep, position: Int, documentId: String) {
+    override suspend fun saveStoryStep(storyStep: StoryStep, position: Double, documentId: String) {
         documentsMap["root"]?.let { document ->
             val newContent = document.content + (position to storyStep)
             documentsMap["root"] = document.copy(content = newContent)
@@ -97,7 +97,7 @@ class InMemoryDocumentRepository : DocumentRepository {
         documentsMap["root"]?.let { document ->
             val newContent = document.content.toMutableMap()
             steps.forEach { (position, storyStep) ->
-                newContent[position.toInt()] = storyStep
+                newContent[position] = storyStep
             }
             documentsMap["root"] = document.copy(content = newContent)
         }
@@ -106,7 +106,7 @@ class InMemoryDocumentRepository : DocumentRepository {
     override suspend fun deleteStoryStep(storyStepId: String, documentId: String) {
         documentsMap["root"]?.let { document ->
             val newContent = document.content.filterNot { (_, step) -> step.id == storyStepId }
-                .values.mapIndexed { index, step -> index to step }.toMap()
+                .values.mapIndexed { index, step -> index.toDouble() to step }.toMap()
             documentsMap["root"] = document.copy(content = newContent)
         }
     }
@@ -129,7 +129,7 @@ class InMemoryDocumentRepository : DocumentRepository {
     override suspend fun moveDocumentsToWorkspace(oldUserId: String, newUserId: String) {
     }
 
-    override suspend fun updateStoryStep(storyStep: StoryStep, position: Int, documentId: String) {
+    override suspend fun updateStoryStep(storyStep: StoryStep, position: Double, documentId: String) {
     }
 
     override suspend fun search(query: String, workspaceId: String): List<Document> =
