+++ writeopia_ui/src/commonMain/kotlin/io/writeopia/ui/drawer/content/DesktopTextItemDrawer.kt
@@ -41,10 +41,10 @@ class DesktopTextItemDrawer(
     private val modifier: Modifier,
     private val customBackgroundColor: Color,
     private val clickable: Boolean,
-    private val onSelected: (Boolean, Int) -> Unit,
+    private val onSelected: (Boolean, Double) -> Unit,
     private val dragIconWidth: Dp,
     private val config: DrawConfig,
-    private val onDragHover: (Int) -> Unit,
+    private val onDragHover: (Double) -> Unit,
     private val onDragStart: () -> Unit,
     private val onDragStop: () -> Unit,
     private val moveRequest: (Action.Move) -> Unit,
@@ -52,7 +52,7 @@ class DesktopTextItemDrawer(
     private val endContent: @Composable ((StoryStep, DrawInfo, Boolean) -> Unit)? = null,
     private val isDesktop: Boolean,
     private val enabled: Boolean,
-    private val receiveExternalFile: (List<ExternalFile>, Int) -> Unit,
+    private val receiveExternalFile: (List<ExternalFile>, Double) -> Unit,
     private val messageDrawer: @Composable RowScope.() -> SimpleTextDrawer,
     private val tagDecoration: TagDecoration = DefaultTagDecoration,
     private val paddingValues: PaddingValues = PaddingValues(0.dp),
@@ -104,7 +104,7 @@ class DesktopTextItemDrawer(
                 when (inBound) {
                     InBounds.OUTSIDE -> {}
                     InBounds.INSIDE_UP -> {
-                        val position = drawInfo.position - 1
+                        val position = drawInfo.position - 1.0
                         handleDrag(position, data)
                     }
 
@@ -181,7 +181,7 @@ class DesktopTextItemDrawer(
         }
     }
 
-    private fun handleDrag(position: Int, data: DropInfo?) {
+    private fun handleDrag(position: Double, data: DropInfo?) {
         onDragHover(position)
 
         if (data != null) {
